authProject/auth_requirements.txt
Django==4.2.11
psycopg2-binary==2.9.9
djangorestframework==3.15.1
django-filter==24.2
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
django-cors-headers==4.3.1
drf-nested-routers==0.93.4
pytest-django==4.5.2
uwsgi==2.0.25.1
gunicorn==22.0.0

